Add missing messages to the HostControlDispatcher handler
While working on the host I occasionally see "Unknown control
message received" in the log file. I would like to know which
control message was received but it's not simple since the
ControlMessage contains a dozen or so optional fields and does
not use oneof.

I've added some logging statements to the if tree for the current
set of ControlMessages. The unhandled messages will log out which
message they are or 'unknown' in case we add a new message to the
proto w/o adding it to the HostControlDispatcher.

Change-Id: Iaacc14e84b5e9f6d192cf614015739818714307a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5368320
Auto-Submit: Joe Downing <joedow@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272415}