Avoid Apps Collections to show again after dismissed
Add a local value to the controller to store if the apps collections
view was ever dismissed during the session.  If the app list presenter
needs to check if the apps collections view should be shown, verify
first if the apps collections was previously dismissed.

Bug: b:325640336
Change-Id: I0fe0846cef0325415eaa33dc348a914de506ff18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5352453
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272436}