[Android] Trigger clicks on the root view.
Instead of dispatching the MotionEvent on the passed view
directly, this CL changes to logic to dispatch the MotionEvent
on the root view with recomputed coordinates. This is how touch
events are dispatched in Android, that's why it's correct to do
it this way.

For other views, we'll set the motion event interceptors. This is
why we need to dispatch MotionEvent on the root view.

Bug: b/328278717
Change-Id: Ia7fd10406443395c528d276a220c2296c68db0f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5370195
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1272638}