Implement syncer::SyncMetadataStore for PlusAddressTable
This way, PlusAddressTable can be used as the metadata store for
PLUS_ADDRESS. The functions are implemented using the tables added in
crrev.com/c/5347717.

Besides the interface, a function GetAllSyncMetadata() is added, which
populates a syncer::MetadataBatch with all stored sync metadata. This is
needed to call ModelTypeChangeProcessor::ModelReadyToSync() in the
initialisation of PlusAddressSyncBridge.

Bug: b/322147254
Change-Id: I2908c2f161fa98c2a2d6d24fce882b7294650352
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5349149
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1272644}