[//gpu] Give ClientSI correct format for per-plane sampling
When a SharedImage is created from a {GPUMemoryBuffer, plane} pair,
the associated ClientSharedImage currently always gets the format of the
buffer as a whole. However, when using per-plane sampling, it should
actually have the format of the *plane* (as the service side will). This
CL makes that change.

In addition to general correctness, this will be critical for allowing
ClientSharedImage::GetTextureTarget() to correctly detect when external
sampling is being used based only on its format.

Bug: 41494843
Change-Id: I8525a1ea821dedc1ea341da2412701af17f07cf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5369800
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272652}