[UNOp2][Android] Enable bookmark sync only for completed signin
When using the sign-in promo from the bookmark manager without the UNO2
flag enabled, the bookmark syncing should be enable only if the sign-in
is successful.

This CL moves the sync enabling logic from the sign-in flow completion
callback to the sign-in completion callback, to fix the crash which
occurs when dismissing the sign-in bottom sheet from the bookmarks
sign-in promo.

Bug: 41493773
Change-Id: Ibb73a9026bff70065c3f08e59a6a972afaa0e28d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5370014
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Auto-Submit: Menghan Yang <myuu@google.com>
Commit-Queue: Menghan Yang <myuu@google.com>
Cr-Commit-Position: refs/heads/main@{#1272675}