[UPM] Introduce IsGmsCoreUpdateRequired method for blocking UI M4
The method will be used to identify whether to show the blocking warning (or error message) instead of password manager UI in various places.

P. S. The method
`PasswordStoreAndroidBackendBridgeHelper::CanCreateBackend()` can't be
accessed from components\ because it would be a circular
dependency. To avoid this, I introduced the const
`kDefaultAccountUpmMinGmsVersion`, which allows to check the min Gms
version for syncing users. Removal of
`PasswordManagerCreationCheckUtil.isMinGmsCoreVersionAvailable()` from
clank and rewriting of
`PasswordStoreAndroidBackendBridgeHelper::CanCreateBackend()` will be
done in a follow-up CL.

Bug: 321217920
Change-Id: Ie45350e529b59daee270fb0cf2cd423f79ef19d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5356533
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1272705}