OOBE: Add metrics client ID getter in test_api
Add a getter for the metrics client ID to introduce tast tests
checking for the consistency of the ID during the OOBE flow.

In the early part of OOBE flow, the initial ID is stored in the
preference `metrics::prefs::kMetricsProvisionalClientID` and only
promoted after the acceptance of consolidated consent.
Before the promotion, `metrics_service()->GetClientId()` would
return an empty string.

Bug: b:329445169
Change-Id: I68b24618864b82931dbc61252c5cb48b1ded74ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5369916
Reviewed-by: Bohdan Tyshchenko <bohdanty@google.com>
Commit-Queue: Osama Fathy <osamafathy@google.com>
Cr-Commit-Position: refs/heads/main@{#1272746}