Enable the Cronet Experiment Framework by default
Starting from this CL, Cronet will read the HTTP flags file (used for
pushing Cronet experiment flags to devices) by default on
initialization, unless explicitly disabled in the app manifest. This
effectively delivers the Cronet Experiment Framework to production and
makes it possible to use it to run real experiments on a variety of
first-party and third-party apps running on real devices in the real
world.

Bug: b:300238027
Change-Id: I430184e9e8c6b079de8d7906866e49cbdfbeae65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366025
Reviewed-by: Stefano Duo <stefanoduo@google.com>
Commit-Queue: Etienne Dechamps <edechamps@google.com>
Cr-Commit-Position: refs/heads/main@{#1272755}