Accelerate `UTF16RagelIterator`
Improve caching: Due to the Ragel scanner using two different
UTF16RagelIterator objects, this frequently lead to situations where
we would call the character properties calls redundantly for the same
cursor position. Improve the caching logic to avoid that, by moving
the lookup to dereferencing time.

This leads to speedups of 1.77x vs 1.22x speed-up respectively in the
synthetic SymbolsIterator benchmark [1]:

Without [2], average of 3 runs of [1]:
21.033s vs. 17.337s = 1.21x speed-up

With [2], avg of 3 runs of [1]:
5.154s vs 3.064s = 1.68x speed-up

The difference between with [2]/without [2] is explained by the relative
amount of time that is spent in UTF16RagelIterator, which increases
after the optimizations in [2].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5307975
[2] https://chromium-review.googlesource.com/c/chromium/src/+/5317569

Bug: 41495057, 325814602
Change-Id: Id94ce034bd6b601cae65e354e74123bd9b469f7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5317649
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272756}