Allow saving Mexican addresses with only a municipio but no city
Mexico is special in that it has cities and municipios/demarcaciones
territoriales. Some websites ask for only one of the two and we still
want to offer saving addresses.

This CL maps the "Address needs a city" requirement to "Address needs a
city or municipio" in MX.

Bug: 40275657
Change-Id: I0806fd8f7069acf6a82309faacd9d0834ed45395
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366589
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Dominic Battre <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272764}