WebViewPrerender: Prerender support for AwContentsIoThreadClient
AwContentsIoThreadClient manages the correspondence of RFHs and Java IO
threads. Before this CL, AwContentsIoThreadClient assumed that the
FrameTreeNode associated with an RFH never change in its lifetime,
which doesn't hold in prerender activation.

This CL makes AwContentsIoThreadClient prerender-ready by adding
RenderFrameHost/FrameTree swap handling.

Note that the change is behind a feature flag kWebViewPrerender2 and
does nothing if it is disabled.

This CL is necessary for the first test of Prerender for WebView. We'll
add a test for back navigation after [2] landed.

[1] https://docs.google.com/document/d/1Pnf6c7jF04RpA6m_lStTKofLAb-xBoc65H_BKhy-KRE
[2] https://crrev.com/c/5269709

Credit: This CL was written by kenoss@ (not nhiroki@)

Fixed: b/41492762
Change-Id: I02b9e293f69a8ed0f9322e3087a4aa05041228ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5316140
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272806}