Don't send touch moves to EventForwarder observers.
This is part of Input Vizard
project(https://docs.google.com/document/d/1mcydbkgFCO_TT9NuFE962L8PLJWT2XOfXUAPO88VuKE)
where we are trying to move input to GPU process. With input coming on
GPU process we would need to send back the inputs back to browser for
the observers which cannot be moved to GPU itself. We are trying to
minimize sending inputs back to Browser process to minimize the IPCs we
make.

With this change we are stopping sending touch moves to observers early
in project so that we get to know about any breaking changes early i.e.
if we need to send touch moves back to browser.

This shouldn't break the only existing observer of EventForwarder touch
events(AttributionInputEventTrackerAndroid) since it only cares about
events leading to navigations and touch moves shouldn't be causing any
navigations.

Bug: b/328602561
Change-Id: I77eb877da5b6c11fc3d9241bc91c3bd842d0f778
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5355053
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Kartar Singh <kartarsingh@google.com>
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272807}