[//media] RenderableGMBVideoFramePool get texture target from ClientSI
This CL converts RenderableGMBVideoFramePool away from overriding the
texture target to use on MacOS only to call
ClientSharedImage::GetTextureTargetForOverlays(), which has identical
semantics to the current flow:

* Returns GetPlatformSpecificTextureTarget() on Mac
* Return GL_TEXTURE_2D otherwise

This is a step on the road to the eventual goal of having all client
texture target computation behind a universal
ClientSharedImage::GetTextureTarget() function.

This CL also removes a now-obsolete TODO from this code.

Note: There is existing weirdness in this code with the texture target
of the MailboxHolders sometimes getting set and sometimes not in the
case where the ClientSharedImage was not successfully created. This
CL does not attempt to fix that code, although we should separately.

Bug: 41494843
Change-Id: I7d08b978c4cfdd68e6658f4457929f5e4eb6cf20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366414
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272824}