[Signin][Android] Remove FirstRunFlowSequencerDelegate.isSyncAllowed
Checks in isSyncAllowed are redundant, as FRE will only show the sync
opt-in page (or the history sync opt-in page) to users who have signed
in on the first FRE page. This CL removes isSyncAllowed as well as some
helper methods from FirstRunUtils that are no longer needed.

This CL is a pre-requisite to https://crrev.com/c/5365142 (as
FirstRunUtilsTest was failing in that CL).

Bug: 40890215
Change-Id: I6c0057271ea9cc391830b055fab1e372dc76d674
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366207
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Tanmoy Mollik <triploblastic@google.com>
Cr-Commit-Position: refs/heads/main@{#1272828}