[Cast Convergence] Remove is_castos = true builders from try jobs.
This is part 1/3 of a multi-part change to update Cast Linux builders
in Chromium.  The is_castos flag is no longer supported in upstream
Chrome.

The following builders will be turned down:
- Cast Linux ARM64 / linux-arm64-castos
- Cast Linux Debug / linux-x64-castos-dbg
- Cast Linux / linux-x64-castos

They will be replaced with roughly the following:

- linux-arm64-cast-rel (for validating builds pushed to devices)
- linux-x64-cast-dbg (for validating that desktop builds work)

Some, but not all of the test suites for the old builders will be
re-used for the new builders.  The unused suites will be deleted in a
followup CL.

Android and Fuchsia cast builders remain unaffected.

Bug: 1516671
Change-Id: Ifdff72fd388975b073c928a66d4ef50d9eb5e294
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5368254
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Mark Foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272837}