[Default Nav Transition] Implement the UX for BeforeUnload handler
The UX is: as long as the browser sends out the BeforeUnload IPC to the
renderer, we play the cancel animation to bring back the active page,
even if the renderer decides to not show the prompt at all (e.g., if
the page doesn't have a sticky user activation).

In real world testing, the IPC exchange is so fast (in the case where
the renderer doesn't show the prompt and acks to proceed) that the
cancel animation never gets to play a single frame (even on a debug
build). This is close the the ideal UX mentioned below.

Ideally we want to only play the cancel animation if the renderer
decides to show the prompt. While this is a better UX, it makes the
impl complicated as we need to decide what to show after the user lifts
the finger and before the renderer asks the browser to draw the prompt.
(presumably the invoke animation, then plays the cancel animation when
the renderer acks to show the prompt).

Alternatively, we could let the browser not send out the BeforeUnload
message if we know that the renderer for sure will always ack with
proceed. RFHI::HasStickyUserActivation already has the information,
though the source of truth is passed from the renderer. However we
don't really know the complication of changing the BeforeUnload flow
so this current impl is the best option for now.

Bug: 1421009
Low-Coverage-Reason: TRIVIAL_CHANGE
Change-Id: I993faba663075f5519321d69c26afa2b26404974
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5325441
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: William Liu <liuwilliam@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272861}