Log AndroidTelemetryService referrer chain metrics at ping generation
The AndroidTelemetryService currently observes all downloads, then only
does anything with a small fraction of them. This means that
https://crrev.com/c/5318781 began computing referrer chains for all
downloads, and logging metrics for all downloads. This is not the intent
of the metrics, so instead persist the result of referrer chain
computation so we can log it only when we decide to send a ping.

Bug: 324293877
Change-Id: I661f792f14e8b3c852b37b8e3661ad63cea9e02e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5369575
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272900}