[Nearby Presence] Migrate NearbyPresenceService::PresenceDevice
`NearbyPresenceService::PresenceDevice` (NPS::PD) was originally created
before the library representation of PresenceDevice was completed. This
change removes the usage of NPS::PD and replaces it with the Nearby
library's implementation.

Note that doing so removes the RSSI and Stable Device ID information, as
that's currently not being tracked in the library implementation. A
future CL will either re-introduce it into the library if that data is
still needed OR remove those fields from the mojom::PresenceDevice
if they are no longer needed.

Fixed: b/327278661
Test: Unit tests updated
Change-Id: I95f37d5d5abbad59f6a172fffe60d8f01d14e514
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5329932
Commit-Queue: Crisrael Lucero <crisrael@google.com>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Juliet Lévesque <julietlevesque@google.com>
Cr-Commit-Position: refs/heads/main@{#1272907}