webui: detect when the preloaded contents is ready to show
A top chrome WebUI page notifies its embedder when it is ready to be
shown by calling MojoBubbleWebUIController::Embedder::ShowUI(). This
patch captures this signal in the preload manager, when the page is not
yet attached to any views.

WebUIPreloadManager::MakeContents() now returns a struct that contains
the preloaded WebContents and an "is_ready_to_show" boolean. This will
later be used in WebUIBubbleManager to skip page reloading if the page
is already ready to show.

Bug: 325928324
Change-Id: I3e89a5159db04de54d3c1de0a418c027f5c1dfcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366285
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Keren Zhu <kerenzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272924}