Roll src/ios/third_party/earl_grey2/src/ 92f42b46a..6acfcc10d (1 commit)
https://chromium.googlesource.com/external/github.com/google/EarlGrey.git/+log/92f42b46a340..6acfcc10d3cb

$ git log 92f42b46a..6acfcc10d --date=short --no-merges --format='%ad %ae %s'
2024-02-16 nishantdalmia Internal Cleanup

Created with:
  roll-dep src/ios/third_party/earl_grey2/src

selectElementWithMatcher now requires that its return value be used. See:
https://chromium.googlesource.com/external/github.com/google/EarlGrey.git/+/6acfcc10d3cbbd19b428cf5e8fa26d86b383a1ea%5E%21/#F0
There are 3 tests in chromium which do nothing with its return value and
have caused the build to fail. This CL fixes two of them and assigns
a TODO to the test owner of the third (since the fix did not appear
obvious).

Originally submitted https://crrev.com/c/5344217, but something in that CL range caused EG test failures on the iphone-device builder. This CL rolls only the change that broke the 3 EG tests listed. We will fix the failures on iphone-device in a subsequent roll.

Bug: 328267991,328271981
Change-Id: I3d8f131b207cdf4401da6dd630e5b7f906b6d5b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366744
Reviewed-by: Lindsay Pasricha <lindsayw@chromium.org>
Commit-Queue: Will Yeager <wyeager@google.com>
Reviewed-by: Adam Arcaro <adamta@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272940}